chore(testworkflows): TKC-1763 - decoupling spawn implementation
